+++ MemeFyMe/app/src/main/java/lighterletter/c4q/nyc/memefymeapp/MainActivity.java
@@ -1,25 +1,16 @@
 package lighterletter.c4q.nyc.memefymeapp;
 
 import android.content.Intent;
-
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
 import android.support.v7.app.ActionBarActivity;
-
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-
 import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-
-import android.support.v7.app.ActionBarActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
 
 
 
@@ -107,24 +98,7 @@ public class MainActivity extends ActionBarActivity {
         }
 
 
-
-        //setupEvents();
-
     }
-    private void setupEvents() {
-        Button shareTextButton = (Button)findViewById(R.id.main_share_button);
-        shareTextButton.setOnClickListener(new View.OnClickListener() {
 
-            @Override
-            public void onClick(View v) {
-                Intent picIntent = new Intent(getApplicationContext(), SharePictureActivity.class);
-                startActivity(picIntent);
-
-            }
-        });
-
-
-
-    }
 }
