+++ MemeFyMe/app/src/main/res/layout/activity_share.xml
@@ -4,18 +4,13 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="lighterletter.c4q.nyc.memefymeapp.VanillaMemeSampleActivity">
+    tools:context="lighterletter.c4q.nyc.memefymeapp.ShareActivity">
 
     <LinearLayout
         android:orientation="vertical"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent">
 
-        <ImageView
-            android:layout_width="fill_parent"
-            android:layout_height="fill_parent"
-            android:layout_weight="2"
-            android:id="@+id/original" />
 
         <ImageView
             android:layout_width="fill_parent"
