--- MemeFyMe/app/src/main/res/layout/activity_share_picture.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <TextView android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/share_picture_description" />
-
-    <LinearLayout android:layout_width="200dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:orientation="vertical">
-
-        <ImageView android:id="@+id/share_picture_thumbnail"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:contentDescription="@string/share_picture_thumbnail_content_description" />
-
-        <Button android:id="@+id/share_picture_camera_button"
-            android:layout_width="fill_parent"
-            android:layout_height="80dp"
-            android:text="@string/share_picture_camera_button_text" />
-
-    </LinearLayout>
-
-    <Button android:id="@+id/share_picture_share_button"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/title_share" />
-
-</LinearLayout>
