+++ MemeFyMe/app/src/main/res/layout/fragment_demotivational.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/black"
-    tools:context="lighterletter.c4q.nyc.memefymeapp.Demotivational">
+    tools:context="lighterletter.c4q.nyc.memefymeapp.DemotivationalFragment">
 
     <FrameLayout
         android:id="@+id/demotivationalView"
