+++ MemeFyMe/app/src/main/res/layout/fragment_vanilla.xml
@@ -1,7 +1,7 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@android:color/black" tools:context="lighterletter.c4q.nyc.memefymeapp.Vanilla">
+    android:background="@android:color/black" tools:context="lighterletter.c4q.nyc.memefymeapp.VanillaFragment">
 
 
     <FrameLayout
