+++ MemeFyMe/app/src/main/res/menu/menu_share.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="lighterletter.c4q.nyc.memefymeapp.VanillaMemeSampleActivity">
+    tools:context="lighterletter.c4q.nyc.memefymeapp.ShareActivity">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
